+++ src/common/components/ProgramBudgetList.jsx
@@ -1,0 +1,129 @@
+import React, { useState, useEffect } from 'react';
+import styled from 'styled-components';
+
+const BudgetListContainer = styled.div`
+  margin: 2rem auto;
+  width: 100%;
+  max-width: 1000px;
+  background-color: white;
+  border-radius: 12px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.08);
+  padding: 1.5rem;
+
+  h2 {
+    margin-top: 0;
+    margin-bottom: 1.25rem;
+    color: #333;
+    font-weight: 600;
+    text-align: center;
+    padding-bottom: 0.75rem;
+    border-bottom: 1px solid #eee;
+  }
+`;
+
+const BudgetGrid = styled.div`
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 1.25rem;
+`;
+
+const BudgetCard = styled.div`
+  padding: 1rem;
+  background-color: #f9fafb;
+  border-radius: 8px;
+  transition:
+    transform 0.2s ease,
+    box-shadow 0.2s ease;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  .program-title {
+    font-weight: 500;
+    margin-bottom: 0.5rem;
+    font-size: 0.9rem;
+    color: #555;
+    text-align: center;
+  }
+
+  .budget-amount {
+    font-weight: 700;
+    font-size: 1.25rem;
+    color: var(--green-9, #2e7d32);
+  }
+`;
+
+const StatusMessage = styled.div`
+  margin-top: 1rem;
+  color: ${(props) => (props.error ? 'red' : '#666')};
+  font-style: italic;
+`;
+
+export default function ProgramBudgetList() {
+  const [programBudgets, setProgramBudgets] = useState([]);
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState(null);
+
+  useEffect(() => {
+    const fetchProgramBudgets = async () => {
+      try {
+        setLoading(true);
+        setError(null);
+
+        const response = await fetch(
+          `${process.env.REACT_APP_BACKEND_URL}/budget`
+        );
+
+        if (!response.ok) {
+          throw new Error(`Could not get budgets! Status: ${response.status}`);
+        }
+
+        const programs = await response.json();
+        setProgramBudgets(programs);
+      } catch (err) {
+        setError(`Error fetching budgets: ${err.message}`);
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchProgramBudgets();
+  }, []);
+
+  if (loading) {
+    return <StatusMessage>Loading program budgets...</StatusMessage>;
+  }
+
+  if (error) {
+    return <StatusMessage error>{error}</StatusMessage>;
+  }
+
+  if (programBudgets.length === 0) {
+    return <StatusMessage>No program budgets available</StatusMessage>;
+  }
+
+  return (
+    <BudgetListContainer>
+      <h2>Program Budget</h2>
+      <BudgetGrid>
+        {programBudgets.map((program) => {
+          // Helper function to capitalize first letter
+          const capitalizeFirstLetter = (string) => {
+            if (!string) return 'Unknown Program';
+            return string.charAt(0).toUpperCase() + string.slice(1);
+          };
+          return (
+            <BudgetCard key={program.program_id}>
+              <div className='program-title'>
+                {capitalizeFirstLetter(program.program_title)}
+              </div>
+              <div className='budget-amount'>
+                ${parseFloat(program.program_budget).toFixed(2)}
+              </div>
+            </BudgetCard>
+          );
+        })}
+      </BudgetGrid>
+    </BudgetListContainer>
+  );
+}
